+++ src/index.ts
@@ -403,6 +403,19 @@ class WrapBlox {
 		return response.ok;
 	}
 	
+	/**
+	 * 
+	 * @param id group ID
+	 * @param userId user ID to decline
+	 * @returns was the request successful
+	 */
+	
+	async declineJoinRequest(id : number, userId : number) : Promise<boolean> {
+		if (!this.cookie) return false;
+		const response = await this.delete("groups", `groups/${id}/join-requests/users/${userId}`);
+		return response.ok;
+	}
+